AI_Radiology_Detection_Project/ai-radiology-detection/backend/requirements-report.txt
flask==3.0.0
flask-cors==4.0.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
PyJWT==2.8.0
reportlab==4.1.0
